+++ NEWS.md
@@ -1,3 +1,9 @@
+# seq2vcf 0.1.1
+
+* Fixed errors
+* Realised insertions and deletion are not taken into account by snp-align
+* added remove gaps function (from the DECIPHER) package to remove gaps only from the first sequence
+
 # seq2vcf 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.